Moved Scenehandler and TimelineParser to core, added CreatorMain and ViewerMain
