+++ Slideshow-Creator/src/creator/CreatorMain.java
@@ -1,0 +1,5 @@
+package creator;
+
+public class CreatorMain {
+
+}
