+++ Slideshow-Creator/src/viewer/ViewerMain.java
@@ -1,0 +1,5 @@
+package creator;
+
+public class ViewerMain {
+
+}
